fix: 打碼與充值量 type型別錯誤

+++ logic/room.go
@@ -26,10 +26,10 @@ type Limit struct {
 	Day int `json:"day" binding:"max=31"`
 
 	// 儲值金額
-	Deposit int `json:"deposit"`
+	Deposit float64 `json:"deposit"`
 
 	// 打碼量
-	Dml int `json:"dml"`
+	Dml float64 `json:"dml"`
 }
 
 func (l *Logic) CreateRoom(r Room) (string, error) {
